Fix python
